requirements.txt
streamlit
pillow
torch
torchvision
numpy
matplotlib
tqdm
torchmetrics
torchinfo